v 2023.3
This version fixes an issue when a target could be shown as not vulnerable when acceptableDifferenceLengthBetweenResponses was not set to -1. Setting acceptableDifferenceLengthBetweenResponses to -1 can also create false positives in some cases so this needed fixing. The issue was in a Regular Expression which was used to clean dynamic contents from responses to make them ready for a comparison.

As a result, two new configs have been also added to handle edge cases when dealing with specific targets: ignoreHeaderLevel & ignoreBodyLevel